refactor packfile.core into packfile.parser
